requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
psycopg2-binary
requests